убрал ненужный AuthorRestController перенес его функционал в AdminRestController убрал ненужный AuthorDto поправил js и html страницы
